backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8

# LLM and Embeddings
openai==1.3.7
anthropic==0.7.4
langchain==0.0.340
tiktoken==0.5.1

# Vector Store
chromadb==0.4.18
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1